modal functionality works

+++ public/javascript/delete-gift.js
@@ -1,19 +1,28 @@
 async function deleteFormHandler(event) {
-    event.preventDefault();
+	event.preventDefault();
 
-    // retrieves item id
-    const id = document.querySelector(".gift-id").textContent
-    console.log(id)
-    
-    const response = await fetch(`/api/listItems/${id}`, {
-        method: 'DELETE'
-    });
+    const href = event.target.parentElement.getAttribute("href")
+	// retrieves item id
+	const id = event.target.id;
 
-    if (response.ok) {
-        document.location.replace('/dashboard/');
+	const deleteBtn = document.getElementById(id);
+	console.log(deleteBtn);
+
+	if (deleteBtn) {
+		const response = await fetch(`/api/listItems/${id}`, {
+			method: "DELETE",
+		});
+
+		if (response.ok) {
+			document.location.replace('/dashboard/');
+		} else {
+			alert(response.statusText);
+		}
     } else {
-        alert(response.statusText);
+        window.open(href)
     }
 }
 
-document.querySelector('.delete-gift-btn').addEventListener('click', deleteFormHandler);
+document
+	.querySelector("#gifts-container")
+	.addEventListener("click", deleteFormHandler);
